Build a stepped-wedge cluster randomised trial design binary treatment indicator matrix

design_sw builds a binary treatment indicator matrix for a specified stepped-wedge cluster randomised trial (SW-CRT) design, for subsequent use with draw_sw.

design_sw(clusters_per_wave = c(0, 1, 1, 1, 1),
          labels            = c("0", "1"),
          row_names         = 1:sum(clusters_per_wave),
          col_names         = 1:length(clusters_per_wave))

Arguments

clusters_per_wave
...

labels
...

row_names
...

col_names
...

Value

Returns the binary treatment indicator matrix for the implied SW-CRT design. This can be passed to draw_sw.

See also

Author

Michael J Grayling (michael.grayling@newcastle.ac.uk)

Examples

# The default is a 'standard' SW-CRT design, with four clusters and five time
# periods
default_design <- design_sw()
# Specify a more complex design, where all clusters begin in the intervention
# condition, there is an unequal number of clusters per wave, and there are
# two trailing extra time periods with all clusters in the intervention
# condition. Also modify the row and column names
complex_design <- design_sw(clusters_per_wave = c(1, 2, 3, 2, 0, 0),
                            row_names         = paste0("Cluster ", 1:8),
                            col_names         = paste0("Time period ", 1:6))