swcrt: Draw and export stepped-wedge cluster randomised trial design diagrams

swcrt is an R package that helps produce diagrams of stepped-wedge cluster randomised trial (SW-CRT) designs, like those advocated for inclusion when reporting SW-CRTs by Hemming et al. (2018). The produced diagrams can be exported in .docx, .png, and .pptx formats.

Getting started

You can install the latest development version of swcrt from Github with:

devtools::install_github("mjg211/swcrt")

An introductory example of how to make use of the package's core functionality can be found here. More detailed support is available in the package vignette, which can be accessed with vignette("swcrt"). For further help, please email michael.grayling@newcastle.ac.uk.

Details

Package:	swcrt

Type:	Package

Version:	0.1

Date:	2020-09-10

License:	MIT

The package exports two functions:

draw_sw: Draws SW-CRT designs that can be exported in a variety of formats.

design_sw: A helper function to assist in producing standard SW-CRT designs for subsequent use with draw_sw.

References

Hemming K, Taljaard M, McKenzie JE, Hooper R, Copas A, Thompson JA, Dixon-Woods M, Aldcroft A, Doussau A, Grayling M, Kristunas C, Goldstein CE, Campbell MK, Girling A, Eldridge S, Campbell MJ, Lilford RJ, Weijer C, Forbes AB, Grimshaw JM. Reporting of stepped wedge cluster randomised trials: Extension of the CONSORT 2010 statement with explanation and elaboration. BMJ 2018;363,k1614. DOI: 10.1136/bmj.k1614. PMID: 30413417.

Author

Michael J Grayling (michael.grayling@newcastle.ac.uk)