Draw and export stepped-wedge cluster randomised trial design diagrams

...

draw_sw(design         = design_sw(),
        cp_contents    = design,
        row_names      = FALSE,
        col_names      = FALSE,
        colours,
        combine_rows   = FALSE,
        merge_cols     = FALSE,
        xlab           = "Time period",
        ylab           = "Cluster",
        cp_text_colour = "black",
        key_interventions,
        key_time_periods,
        make           = "print",
        filename       = "swcrt")

Arguments

design
...

cp_contents
...

row_names
...

col_names
...

colours
...

combine_rows
...

merge_cols
...

xlab
...

ylab
...

cp_text_colour
...

key_interventions
...

key_time_periods
...

make
...

filename
...

Value

A list containing the following elements:

A flextable in the slot $table_design containing the constructed SW-CRT diagram.

A flextable in the slot $table_interventions containing the constructed key summarising the various intervention states. Will be NULL if input key_interventions is not specified.

A flextable in the slot $table_time_periods containing the constructed key summarising the various time periods. Will be NULL if input key_time_periods is not specified.

A list in the slot $inputs containing each of the input variables.

See also

Author

Michael J Grayling (michael.grayling@newcastle.ac.uk)

Examples

# The default is to draw a 'standard' SW-CRT design, with four clusters and
# five time periods
default_draw   <- draw_sw()
# Specify a more complex design, where all clusters begin in the intervention
# condition, there is an unequal number of clusters per wave, and there are
# two trailing extra time periods with all clusters in the intervention
# condition. Also modify the row and column names
complex_design <- design_sw(clusters_per_wave = c(1, 2, 3, 2, 0, 0),
                            row_names         = paste0("Cluster ", 1:8),
                            col_names         = paste0("Time period ", 1:6))
# Pass this to draw_sw(), producing .docx, .png, and .pptx files, merging
# columns, and combining rows
if (FALSE) {
complex_draw   <- draw_sw(complex_design, combine_rows = TRUE,
                          merge_cols = TRUE,
                          make = c("docx", "png", "pptx", "print"))
}